requirements.txt
absl-py==1.4.0
aiohttp==3.9.0
aiosignal==1.3.1
alembic==1.15.2
aniso8601==10.0.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.12.1
bleach==6.0.0
blinker==1.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.4
contourpy==1.1.1
coverage==7.8.0
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@864913f0e57e87a75c8cc0c7d79ecbd774fc669b
distro==1.8.0
Django==4.2.7
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.13.1
Flask==3.0.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.2
fvcore==0.1.5.post20221221
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.2.1
grpcio==1.58.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.9.0
httpcore==1.0.2
httpx==0.25.1
hydra-core==1.3.2
idna==3.4
image==1.5.33
iniconfig==2.1.0
iopath==0.1.9
ipykernel==6.25.2
ipython==8.15.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
keras==2.13.1
kiwisolver==1.4.5
libclang==16.0.6
limits==5.1.0
lxml==4.9.3
Mako==1.3.10
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==4.0.0
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.0.3
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.3.5
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
ordered-set==4.1.0
overrides==7.4.0
packaging==23.1
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
pluggy==1.5.0
portalocker==2.8.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycocotools==2.0.7
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydot==3.0.2
pydub==0.25.1
Pygments==2.16.1
PyJWT==2.10.1
pyparsing==3.1.1
PyQt6==6.6.0
PyQt6-Qt6==6.6.0
PyQt6-sip==13.6.0
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.1.0
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.10.3
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.2
simpleaudio==1.0.4
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
soupsieve==2.5
SQLAlchemy==2.0.40
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.2.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
treelib==1.7.0
typing_extensions==4.13.2
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wavio==0.0.8
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
Werkzeug==3.0.1
wrapt==1.15.0
yacs==0.1.8
yarl==1.9.3
